#7.4 Protecting Video Routes

+++ controllers/videoController.js
@@ -59,7 +59,7 @@ export const videoDetail = async (req, res) => {
       res.redirect(routes.home);
     } else {
       const video = await Video.findById(id).populate("creator");
-      console.log(video);
+      // console.log(video);
       // console.log(req.params);
       res.render("videoDetail", { pageTitle: video.title, video });
     }
@@ -75,7 +75,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     console.log(error);
     res.redirect(routes.home);
@@ -101,7 +105,12 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
-    await Video.findOneAndDelete({ _id: id });
+    const video = await Video.findById(id);
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndDelete({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
